refactor: split monolithic e2e test file into focused modules
Break tests/e2e/cli.test.mjs (3163 lines, 111 tests) into:
- helpers.mjs: shared setup, shim utilities, before/after hooks
- parsing.test.mjs: arg parsing, cosign skip, env-file, cloud mode, security (20 tests)
- adapters.test.mjs: pi/opencode/hermes adapters, agent selection, image tags (17 tests)
- persistence.test.mjs: ephemeral, interactive, CWD normalization, mise, XDG (15 tests)
- mounts.test.mjs: skills, volumes, context files, XDG config (45 tests)
- runtime.test.mjs: apple container runtime, cosign cache (14 tests)

All 111 tests pass. Zero behavioral changes.
