+++ package.json
@@ -5,7 +5,9 @@
   "bin": {
     "harness": "bin/harness.js"
   },
-  "files": ["bin/"],
+  "files": [
+    "bin/"
+  ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/boldblackai/harness.git"
@@ -25,8 +27,8 @@
     "lint:docker": "hadolint Dockerfile Dockerfile.opencode Dockerfile.hermes",
     "lint:actions": "actionlint",
     "format": "biome format --write .",
-    "test:e2e:cli": "node --test tests/e2e/cli.test.mjs",
-    "test:e2e": "node --test tests/e2e/cli.test.mjs",
+    "test:e2e:cli": "node --test tests/e2e/*.test.mjs",
+    "test:e2e": "node --test tests/e2e/*.test.mjs",
     "test:coverage": "node --experimental-test-coverage --test-coverage-include=\"src/**\" --test-coverage-lines=80 --test-coverage-functions=80 --test-coverage-branches=80 tests/e2e/cli.test.mjs"
   },
   "dependencies": {
